+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This version also includes some major internal modifications on the Geo module.
 
 ## 0.8.9-instant
 * Time is now based on Instant rather than LocalDateTime, TimeZone is managed.
+* Time module is deprecated.
+* Added new number locales (es_BO, en_IE, en_IN, ...) and fix some issues with old ones
+* Added new time locales (fa_IR, nb_NO, zh_TW, ...) and fix some issues with old ones
 
 ## 0.8.9
 > Published 20 Nov 2020